refactor : model패키지 내부 구조 변경
