+++ src/main/java/model/products/ProductBox.java
@@ -1,10 +1,12 @@
-package model;
+package model.products;
 
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
+
+import model.products.Product;
 
 public class ProductBox {
 	private static final String COMMA = ",";
